Move heater helpers into inventory module
