+++ custom_components/termoweb/coordinator.py
@@ -21,11 +21,12 @@ from .const import HTR_ENERGY_UPDATE_INTERVAL, MIN_POLL_INTERVAL
 from .inventory import (
     Node,
     _existing_nodes_map,
+    build_heater_address_map,
     build_node_inventory,
+    normalize_heater_addresses,
     normalize_node_addr,
     normalize_node_type,
 )
-from .nodes import build_heater_address_map, normalize_heater_addresses
 from .utils import float_or_none
 
 _LOGGER = logging.getLogger(__name__)
