+++ custom_components/termoweb/energy.py
@@ -18,12 +18,12 @@ from .api import RESTClient
 from .const import DOMAIN
 from .identifiers import build_heater_energy_unique_id
 from .installation import ensure_snapshot
-from .nodes import (
+from .inventory import (
     build_heater_address_map,
-    ensure_node_inventory,
     normalize_heater_addresses,
     parse_heater_energy_unique_id,
 )
+from .nodes import ensure_node_inventory
 from .throttle import (
     MonotonicRateLimiter,
     default_samples_rate_limit_state,
