+++ custom_components/termoweb/installation.py
@@ -6,8 +6,12 @@ from collections.abc import Callable, Iterable, Sequence
 from typing import Any
 
 from .heater_inventory import HeaterInventoryDetails, build_heater_inventory_details
-from .inventory import Node, build_node_inventory
-from .nodes import heater_sample_subscription_targets, normalize_heater_addresses
+from .inventory import (
+    Node,
+    build_node_inventory,
+    heater_sample_subscription_targets,
+    normalize_heater_addresses,
+)
 
 
 class InstallationSnapshot:
