+++ custom_components/termoweb/nodes.py
@@ -6,18 +6,17 @@ from collections.abc import Iterable, Mapping, MutableMapping
 import logging
 from typing import Any, cast
 
-from .const import DOMAIN
 from .inventory import (
+    HEATER_NODE_TYPES,
     Node,
+    addresses_by_node_type,
     build_node_inventory,
+    normalize_heater_addresses,
     normalize_node_addr,
     normalize_node_type,
 )
 
 _LOGGER = logging.getLogger(__name__)
-
-
-HEATER_NODE_TYPES: frozenset[str] = frozenset({"htr", "acm"})
 
 
 def ensure_node_inventory(
@@ -78,153 +77,6 @@ def ensure_node_inventory(
     return []
 
 
-def parse_heater_energy_unique_id(unique_id: str) -> tuple[str, str, str] | None:
-    """Parse a heater energy sensor unique ID into its components."""
-
-    if not isinstance(unique_id, str):
-        return None
-    stripped = unique_id.strip()
-    if not stripped or not stripped.startswith(f"{DOMAIN}:"):
-        return None
-    try:
-        domain, dev, node, address, metric = stripped.split(":", 4)
-    except ValueError:
-        return None
-    if domain != DOMAIN or metric != "energy":
-        return None
-    if not dev or not node or not address:
-        return None
-    return dev, node, address
-
-
-def addresses_by_node_type(
-    nodes: Iterable[Node],
-    *,
-    known_types: Iterable[str] | None = None,
-) -> tuple[dict[str, list[str]], set[str]]:
-    """Return mapping of node type to address list, tracking unknown types."""
-
-    known: set[str] | None = None
-    if known_types is not None:
-        known = {normalize_node_type(node_type) for node_type in known_types if node_type}
-
-    result: dict[str, list[str]] = {}
-    seen: dict[str, set[str]] = {}
-    unknown: set[str] = set()
-
-    for node in nodes:
-        node_type = normalize_node_type(getattr(node, "type", ""))
-        if not node_type:
-            continue
-        addr = normalize_node_addr(getattr(node, "addr", ""))
-        if not addr:
-            continue
-        type_seen = seen.setdefault(node_type, set())
-        if addr in type_seen:
-            continue
-        type_seen.add(addr)
-        result.setdefault(node_type, []).append(addr)
-        if known is not None and node_type not in known:
-            unknown.add(node_type)
-
-    return result, unknown
-
-
-def build_heater_address_map(
-    nodes: Iterable[Any],
-    *,
-    heater_types: Iterable[str] | None = None,
-) -> tuple[dict[str, list[str]], dict[str, set[str]]]:
-    """Return mapping of heater node types to addresses and reverse lookup."""
-
-    allowed_types: set[str]
-    if heater_types is None:
-        allowed_types = set(HEATER_NODE_TYPES)
-    else:
-        allowed_types = {
-            normalize_node_type(node_type, use_default_when_falsey=True)
-            for node_type in heater_types
-            if normalize_node_type(node_type, use_default_when_falsey=True)
-        }  # pragma: no cover - exercised indirectly in integration
-
-    if not allowed_types:
-        return {}, {}  # pragma: no cover - defensive
-
-    by_type_raw, _ = addresses_by_node_type(
-        nodes,
-        known_types=allowed_types,
-    )  # pragma: no cover - exercised via higher level integration tests
-
-    by_type: dict[str, list[str]] = {
-        node_type: list(addresses)
-        for node_type, addresses in by_type_raw.items()
-        if node_type in allowed_types and addresses
-    }
-
-    reverse: dict[str, set[str]] = {}
-    for node_type, addresses in by_type.items():
-        for address in addresses:
-            reverse.setdefault(address, set()).add(node_type)
-
-    return by_type, reverse
-
-
-def normalize_heater_addresses(
-    addrs: Iterable[Any] | Mapping[Any, Iterable[Any]] | None,
-) -> tuple[dict[str, list[str]], dict[str, str]]:
-    """Return canonical heater addresses and compatibility aliases."""
-
-    cleaned_map: dict[str, list[str]] = {}
-    compat_aliases: dict[str, str] = {}
-
-    if addrs is None:
-        sources: Iterable[tuple[Any, Iterable[Any] | Any]] = []
-    elif isinstance(addrs, Mapping):
-        sources = addrs.items()
-    else:
-        sources = [("htr", addrs)]
-
-    for raw_type, values in sources:
-        node_type = normalize_node_type(
-            raw_type,
-            use_default_when_falsey=True,
-        )
-        if not node_type:
-            continue
-
-        alias_target: str | None = None
-        if node_type in {"heater", "heaters", "htr"}:
-            alias_target = "htr"
-        if alias_target is not None and node_type != alias_target:
-            compat_aliases[node_type] = alias_target
-            node_type = alias_target
-
-        if node_type not in HEATER_NODE_TYPES:
-            continue
-
-        if isinstance(values, str) or not isinstance(values, Iterable):
-            candidates = [values]
-        else:
-            candidates = list(values)
-
-        bucket = cleaned_map.setdefault(node_type, [])
-        seen: set[str] = set(bucket)
-        for candidate in candidates:
-            addr = normalize_node_addr(
-                candidate,
-                use_default_when_falsey=True,
-            )
-            if not addr or addr in seen:
-                continue
-            seen.add(addr)
-            bucket.append(addr)
-
-    cleaned_map.setdefault("htr", [])
-    compat_aliases["htr"] = "htr"
-
-    return cleaned_map, compat_aliases
-
-
 def collect_heater_sample_addresses(
     record: Mapping[str, Any] | None,
     *,
@@ -283,21 +135,3 @@ def collect_heater_sample_addresses(
                 normalized_map["htr"] = normalised
 
     return list(inventory), normalized_map, compat
-
-
-def heater_sample_subscription_targets(
-    addrs: Mapping[Any, Iterable[Any]] | Iterable[Any] | None,
-) -> list[tuple[str, str]]:
-    """Return canonical heater sample subscription target pairs."""
-
-    normalized_map, _ = normalize_heater_addresses(addrs)
-    if not any(normalized_map.values()):
-        return []
-
-    other_types = sorted(node_type for node_type in normalized_map if node_type != "htr")
-    order = ["htr", *other_types]
-    return [
-        (node_type, addr)
-        for node_type in order
-        for addr in normalized_map.get(node_type, []) or []
-    ]
