+++ tests/test_init_setup.py
@@ -24,7 +24,7 @@ from homeassistant.exceptions import ConfigEntryAuthFailed, ConfigEntryNotReady
 from homeassistant.helpers import entity_registry as entity_registry_mod
 
 from custom_components.termoweb.identifiers import build_heater_energy_unique_id
-from custom_components.termoweb.nodes import build_heater_address_map
+from custom_components.termoweb.inventory import build_heater_address_map
 
 
 class FakeWSClient:
